+++ src/linodes/linode/networking/components/index.js
@@ -1,2 +1,4 @@
-export IPTransfer from './IPTransfer';
-export IPSharing from './IPSharing';
+export EditRDNS from './EditRDNS';
+export MoreInfo from './MoreInfo';
+export AddIP from './AddIP';
+export IPList from './IPList';