+++ test/linodes/linode/networking/layouts/SummaryPage.spec.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import sinon from 'sinon';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import { expect } from 'chai';
 
+import { SHOW_MODAL } from '~/actions/modal';
 import { SummaryPage } from '~/linodes/linode/networking/layouts/SummaryPage';
-import { ipv4ns, ipv6ns, ipv6nsSuffix } from '~/constants';
-import { testLinode, testLinode1245 } from '@/data/linodes';
+
+import { testLinode } from '@/data/linodes';
+
 
 describe('linodes/linode/networking/layouts/SummaryPage', () => {
   const sandbox = sinon.sandbox.create();
@@ -16,105 +18,113 @@ describe('linodes/linode/networking/layouts/SummaryPage', () => {
     sandbox.restore();
   });
 
-  describe('IPv4', () => {
-    it('renders address ip and url', async () => {
-      const { ipv4 } = testLinode._ips;
-      const page = shallow(
-        <SummaryPage
-          linode={testLinode}
-        />);
+  const ipv4s = Object.values(testLinode._ips).filter(
+    ip => ip.version === 'ipv4');
+  const ipv6s = Object.values(testLinode._ips).filter(
+    ip => ip.version === 'ipv6');
 
-      const ipv4Section = page.find('#publicIpv4 li');
-
-      const { address, rdns } = ipv4.public[0];
-      expect(ipv4Section.at(0).text()).to.equal(`${address}(${rdns})`);
-    });
-
-    it('renders gateway', async () => {
-      const { ipv4 } = testLinode._ips;
-      const page = shallow(
-        <SummaryPage
-          linode={testLinode}
-        />);
-
-      const gateway = page.find('#ipv4Gateway');
-      expect(gateway.text()).to.equal(ipv4.public[0].gateway);
-    });
-
-    it('renders nameservers', async () => {
-      const page = shallow(
-        <SummaryPage
-          linode={testLinode}
-        />);
-
-      const expectedNameservers = ipv4ns[testLinode.region.id];
-      const nameservers = page.find('#ipv4Nameservers li');
-      nameservers.map((nameserver, i) =>
-        expect(nameserver.text()).to.equal(expectedNameservers[i]));
-    });
-  });
-
-  describe('IPv6', () => {
-    it('renders slaac ip', async () => {
-      const { ipv6 } = testLinode._ips;
-      const page = shallow(
-        <SummaryPage
-          linode={testLinode}
-        />);
-
-      const slaac = page.find('#slaac');
-      expect(slaac.text()).to.equal(`${ipv6.slaac.address} / ${ipv6.slaac.prefix}`);
-    });
-
-    /* TODO: add test when gateways are returned for ipv6 objects */
-    it('renders gateway');
-
-    it('renders nameservers', async () => {
-      const page = shallow(
-        <SummaryPage
-          linode={testLinode}
-        />);
-
-      const expectedNameservers = ipv6nsSuffix.map(
-        prefix => ipv6ns[testLinode.region.id] + prefix);
-      const nameservers = page.find('#ipv6Nameservers li');
-      nameservers.map((nameserver, i) =>
-        expect(nameserver.text()).to.equal(expectedNameservers[i]));
-    });
-
-    /* TODO: add tests when we render global pools */
-    it('renders global pool');
-  });
-
-  it('renders add private ip button when no ipv4s are present', async () => {
-    const page = shallow(
+  it('renders ipv4 addresses', function () {
+    const page = mount(
       <SummaryPage
-        linode={testLinode1245}
-      />);
-
-    const button = page.find('#addPrivateIp');
-    expect(button.length).to.equal(1);
-  });
-
-  it('renders private ip', async () => {
-    const page = shallow(
-      <SummaryPage
+        dispatch={dispatch}
         linode={testLinode}
-      />);
+      />
+    );
 
-    const li = page.find('#privateIpv4 li');
-    const privateIp = testLinode._ips.ipv4.private[0];
+    const rows = page.find('Table').at(0).find('tbody tr');
 
-    expect(li.at(0).text()).to.equal(`${privateIp.address}(${privateIp.rdns})`);
+    expect(rows.length).to.equal(ipv4s.length);
+
+    rows.forEach(function (row) {
+      const columns = row.find('td');
+
+      const address = columns.at(0).text();
+      const ip = testLinode._ips[address];
+
+      expect(columns.at(1).text()).to.equal(ip.rdns);
+      expect(columns.at(2).text().toLowerCase()).to.equal(ip.type);
+    });
   });
 
-  it('renders link_local', async () => {
-    const page = shallow(
+  it('renders ipv6 addresses', function () {
+    const page = mount(
       <SummaryPage
+        dispatch={dispatch}
         linode={testLinode}
-      />);
+      />
+    );
 
-    const li = page.find('#linkLocal li');
-    expect(li.at(0).text()).to.equal(testLinode._ips.ipv6.link_local);
+    const rows = page.find('Table').at(1).find('tbody tr');
+
+    expect(rows.length).to.equal(ipv6s.length);
+
+    rows.forEach(function (row) {
+      const columns = row.find('td');
+
+      const address = columns.at(0).text();
+      const ip = testLinode._ips[address.split('/')[0].trim()];
+
+      expect(columns.at(1).text()).to.equal(ip.rdns || '');
+      expect(columns.at(2).text().toLowerCase()).to.equal(ip.type);
+    });
+  });
+
+  it('opens the add modal on click', function () {
+    const page = mount(
+      <SummaryPage
+        dispatch={dispatch}
+        linode={testLinode}
+      />
+    );
+
+    dispatch.reset();
+    page.find('#add-an-ip-address').props().onClick();
+    expect(dispatch.callCount).to.equal(1);
+    expect(dispatch.firstCall.args[0].type, SHOW_MODAL);
+  });
+
+  it('opens the more info modal on click', function () {
+    const page = mount(
+      <SummaryPage
+        dispatch={dispatch}
+        linode={testLinode}
+      />
+    );
+
+    dispatch.reset();
+    page.find('#more-info').at(0).props()
+        .onClick();
+    expect(dispatch.callCount).to.equal(1);
+    expect(dispatch.firstCall.args[0].type, SHOW_MODAL);
+  });
+
+  it('opens the edit rdns modal on click', function () {
+    const page = mount(
+      <SummaryPage
+        dispatch={dispatch}
+        linode={testLinode}
+      />
+    );
+
+    dispatch.reset();
+    page.find('#edit-rdns').at(0).props()
+        .onMouseDown();
+    expect(dispatch.callCount).to.equal(1);
+    expect(dispatch.firstCall.args[0].type, SHOW_MODAL);
+  });
+
+  it('triggers reset rdns on click', function () {
+    const page = mount(
+      <SummaryPage
+        dispatch={dispatch}
+        linode={testLinode}
+      />
+    );
+
+    dispatch.reset();
+    page.find('#reset-rdns').at(0).props()
+        .onMouseDown();
+    expect(dispatch.callCount).to.equal(1);
+    expect(dispatch.firstCall.args[0].type, SHOW_MODAL);
   });
 });